继续补充遗漏的宗族；防止冲突

+++ noname/library/element/content.js
@@ -11984,11 +11984,11 @@ player.removeVirtualEquip(card);
 					}
 					/* else if ("destroyed" in cardx[j]) {
 						
-					}  */else if (info.destroy) {
+					} else if (info.destroy) {
 						cardx[j].delete();
 						cardx[j].destroyed = info.destroy;
 						continue;
-					}
+					} */
 					if (event.position) {
 						if (_status.discarded) {
 							if (event.position == ui.discardPile) {
